+++ Client/src/Components/Organisms/Calendar/style.tsx
@@ -2,7 +2,6 @@ import styled from "styled-components";
 
 export const Container = styled.div`
   height: 60vh;
-  width: 100%;
+  width: 80%;
   margin: 0 auto;
-  overflow: hidden;
 `;
