+++ Client/src/Components/Templates/Main/style.tsx
@@ -4,13 +4,13 @@ import styled from "styled-components";
 export const GridContainer = styled.div`
   display: grid;
   width: 80vw;
-  height: 270vh;
+  height: 300vh;
   margin: 5vh auto;
   grid-template:
     "스터디 공지사항 공지사항" 1fr
-    "이미지 백준랭킹 프로젝트" 1fr
-    "과방대여 과방대여 자유게시판" 1.2fr
+    "자유게시판 백준랭킹 프로젝트" 1fr
     "사진첩 사진첩 사진첩" 1.2fr
+    "과방대여 과방대여 과방대여" 110vh
     /1fr 1fr 1fr;
   gap: 1vw;
 `;
